requirements.txt
Django==5.0.3
djangorestframework==3.17.1
psycopg2-binary==2.9.11
asgiref==3.11.1
sqlparse==0.5.5
gunicorn==25.3.0
uvicorn==0.44.0
h11==0.16.0
whitenoise==6.12.0
django-environ==0.13.0
django-simple-history==3.11.0
pynetbox==7.3.3
Pillow==10.2.0
packaging==23.2
click==8.3.2
django-debug-toolbar==4.3.0
Werkzeug==3.1.8
MarkupSafe==3.0.3
pysqlite3-binary==0.5.1